requirements.txt
easydict
matplotlib
numpy
pandas
Pillow
PyYAML
scikit_learn
scipy
seaborn
termcolor
timm
torch
torchvision
tqdm
trimesh
umap_learn
wandb
ninja
